+++ src/main/java/vendingmachine/Application.java
@@ -8,6 +8,7 @@ public class Application {
     public static void main(String[] args) {
         VendingMachineGenerator vendingMachineGenerator = new VendingMachineGenerator();
         VendingMachine vendingMachine = vendingMachineGenerator.createVendingMachine();
+        
         VendingMachineController vendingMachineController = new VendingMachineController(vendingMachine);
         vendingMachineController.run();
     }
