+++ src/main/java/vendingmachine/utils/validator/ValidateMoney.java
@@ -1,4 +1,4 @@
-package vendingmachine.validator;
+package vendingmachine.utils.validator;
 
 import vendingmachine.domain.Coin;
 
@@ -6,9 +6,9 @@ public class ValidateMoney {
     public static final int NEGATIVE_ZERO = 0;
     public static final int REMAINDER_ZERO = 0;
 
-    public static final String ERROR_MONEY_NOT_DIGIT = "[ERROR] 금액은 숫자여야 합니다. 다시 입력해주세요.";
-    public static final String ERROR_MONEY_NEGATIVE = "[ERROR] 금액으로 음수를 입력할 수 없습니다. 다시 입력해주세요.";
-    public static final String ERROR_MONEY_UNIT = "[ERROR] 금액을 10원 단위로 입력해야합니다. 다시 입력해주세요.";
+    public static final String ERROR_MONEY_NOT_DIGIT = "\n[ERROR] 금액은 숫자여야 합니다. 다시 입력해주세요.";
+    public static final String ERROR_MONEY_NEGATIVE = "\n[ERROR] 금액으로 음수를 입력할 수 없습니다. 다시 입력해주세요.";
+    public static final String ERROR_MONEY_UNIT = "\n[ERROR] 금액을 10원 단위로 입력해야합니다. 다시 입력해주세요.";
 
     public static boolean isNotValidMoney(String machineMoney) {
         try {
